Moved some projects around, created loopback simulation that passes.
The loopback was initially going to be interactive, but I wasn't able to
easily find a C++ library for non-blocking terminal input, so I ended up
just using a message that is constantly sent out and received back.

Starting to look at building with apio, but I have yet to find out how
to add an extra path to the apio build for yosys.