Changed to comparison in vga_controller and fixed timing issue in test pattern
It turns out that a real VGA monitor will have issues with figuring out the timing if you are
putting out pixels during a blanking portion.  By only having the test pattern output values while
in the active area, I was able to get the full expected color range and frame size as seen in the
simulation.